+++ utils-test-app/utils_test_app/tests/test_esi.py
@@ -16,6 +16,7 @@ from app_utils.esi import (
     fetch_esi_status,
     retry_task_if_esi_is_down,
     retry_task_on_esi_error_and_offline,
+    retry_task_on_esi_issue,
 )
 from app_utils.esi_testing import EsiClientStub, EsiEndpoint, build_http_error
 from app_utils.testing import NoSocketsTestCase
@@ -253,10 +254,50 @@ class TestRetryTaskIfEsiIsDown(NoSocketsTestCase):
         self.assertTrue(kwargs["countdown"])
 
 
+class TestRetryTaskOnEsiErrorAndOffline(NoSocketsTestCase):
+    def test_should_complete_normally_when_no_issue(self):
+        task = Mock(spec=Task)
+        with retry_task_on_esi_error_and_offline(task, "dummy"):
+            pass
+
+    def test_should_retry_when_esi_is_offline_1(self):
+        task = Mock(spec=Task)
+        task.request.retries = 1
+        task.retry.side_effect = CeleryRetry
+        with self.assertRaises(CeleryRetry):
+            with retry_task_on_esi_error_and_offline(task, "dummy"):
+                raise build_http_error(502)
+
+    def test_should_retry_when_esi_is_offline_2(self):
+        task = Mock(spec=Task)
+        task.request.retries = 1
+        task.retry.side_effect = CeleryRetry
+        with self.assertRaises(CeleryRetry):
+            with retry_task_on_esi_error_and_offline(task, "dummy"):
+                raise build_http_error(503)
+
+    def test_should_retry_when_error_limit_is_exceeded(self):
+        task = Mock(spec=Task)
+        task.request.retries = 1
+        task.retry.side_effect = CeleryRetry
+        with self.assertRaises(CeleryRetry):
+            with retry_task_on_esi_error_and_offline(task, "dummy"):
+                raise ESIErrorLimitException(reset=30)
+
+    def test_should_reraise_other_http_errors(self):
+        task = Mock(spec=Task)
+        task.request.retries = 1
+        task.retry.side_effect = CeleryRetry
+
+        with self.assertRaises(HTTPError):
+            with retry_task_on_esi_error_and_offline(task, "dummy"):
+                raise build_http_error(400)
+
+
 class TestRetryTaskOnEsiIssues(NoSocketsTestCase):
     def test_should_complete_normally_when_no_issue(self):
         task = Mock(spec=Task)
-        with retry_task_on_esi_error_and_offline(task):
+        with retry_task_on_esi_issue(task):
             pass
 
     def test_should_retry_when_esi_is_offline_1(self):
@@ -265,41 +306,45 @@ class TestRetryTaskOnEsiIssues(NoSocketsTestCase):
         task.request.retries = 1
         task.retry.side_effect = CeleryRetry
         with self.assertRaises(CeleryRetry):
-            with retry_task_on_esi_error_and_offline(task):
+            with retry_task_on_esi_issue(task):
                 raise build_http_error(502)
 
     def test_should_retry_when_esi_is_offline_2(self):
         task = Mock(spec=Task)
+        task.name = "task_name"
         task.request.retries = 1
         task.retry.side_effect = CeleryRetry
         with self.assertRaises(CeleryRetry):
-            with retry_task_on_esi_error_and_offline(task):
+            with retry_task_on_esi_issue(task):
                 raise build_http_error(503)
 
     def test_should_retry_when_error_limit_is_exceeded(self):
         task = Mock(spec=Task)
+        task.name = "task_name"
         task.request.retries = 1
         task.retry.side_effect = CeleryRetry
         with self.assertRaises(CeleryRetry):
-            with retry_task_on_esi_error_and_offline(task):
+            with retry_task_on_esi_issue(task):
                 raise ESIErrorLimitException(reset=30)
 
     def test_should_retry_when_rate_limit_is_exceeded(self):
         task = Mock(spec=Task)
+        task.name = "task_name"
         task.request.retries = 1
         task.retry.side_effect = CeleryRetry
         bucket = Mock(spec=ESIRateLimitBucket)
         bucket.slug = "bucket-slug"
         bucket.window = 300
         with self.assertRaises(CeleryRetry):
-            with retry_task_on_esi_error_and_offline(task):
+            with retry_task_on_esi_issue(task):
                 raise ESIBucketLimitException(bucket=bucket, reset=30)
 
     def test_should_reraise_other_http_errors(self):
         task = Mock(spec=Task)
+        task.name = "task_name"
         task.request.retries = 1
         task.retry.side_effect = CeleryRetry
 
         with self.assertRaises(HTTPError):
-            with retry_task_on_esi_error_and_offline(task):
+            with retry_task_on_esi_issue(task):
                 raise build_http_error(400)