+++ CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [2.0.0a1] - TBD
+
+### Update notes
+
+This release adds tools for supporting the new ESI rate limits.
+
+New minimum requirements are:
+
+- django-esi 8
+- Alliance Auth 4
+- Python 3.10 (inherited from django-esi 8)
+
+### Added
+
+- `testing.CacheFake`: A fake for replacing Django's cache in tests
+- `testing.reset_celery_once_locks`: Function for deleting celery once locks
+- `esi.retry_task_on_esi_issue`: replaces `esi.retry_task_on_esi_error_and_offline` and adds support for rate limit errors
+
+### Changed
+
+- BREAKING CHANGE: Drops support for Python 3.8, 3.9
+- Adds support for Python 3.12, 3.13
+- `esi.retry_task_on_esi_error_and_offline`: marked as deprecated
+
 ## [1.27.0] - 2025-10-12
 
 ### Added